added new file models.py and added create route
