+++ templates/login.html
@@ -11,8 +11,8 @@
 			<input
 				type="text"
 				id="login"
-				name="username"
-				placeholder="UserName"
+				name="email"
+				placeholder="email"
 			/>
 			<input
 				type="password"
